+++ cgo/cuvs/cagra_c.h
@@ -32,11 +32,11 @@ typedef void* gpu_cagra_c;
 typedef void* gpu_cagra_result_c;
 
 // Constructor for building from dataset
-gpu_cagra_c gpu_cagra_new(const void* dataset_data, uint64_t count_vectors, uint32_t dimension, 
+gpu_cagra_c gpu_cagra_new(const void* dataset_data, uint64_t count_vectors, uint32_t dimension,
                             distance_type_t metric, cagra_build_params_t build_params,
-                            const int* devices, int device_count, uint32_t nthread, 
-                            distribution_mode_t dist_mode, quantization_t qtype, 
-                            const uint32_t* ids, void* errmsg);
+                            const int* devices, int device_count, uint32_t nthread,
+                            distribution_mode_t dist_mode, quantization_t qtype,
+                            const int64_t* ids, void* errmsg);
 
 // Constructor for loading from file
 gpu_cagra_c gpu_cagra_load_file(const char* filename, uint32_t dimension, distance_type_t metric,
@@ -54,17 +54,17 @@ void gpu_cagra_start(gpu_cagra_c index_c, void* errmsg);
 void gpu_cagra_build(gpu_cagra_c index_c, void* errmsg);
 
 // Constructor for an empty index (pre-allocates)
-gpu_cagra_c gpu_cagra_new_empty(uint64_t total_count, uint32_t dimension, distance_type_t metric, 
+gpu_cagra_c gpu_cagra_new_empty(uint64_t total_count, uint32_t dimension, distance_type_t metric,
                                      cagra_build_params_t build_params,
-                                     const int* devices, int device_count, uint32_t nthread, 
-                                     distribution_mode_t dist_mode, quantization_t qtype, 
-                                     const uint32_t* ids, void* errmsg);
+                                     const int* devices, int device_count, uint32_t nthread,
+                                     distribution_mode_t dist_mode, quantization_t qtype,
+                                     const int64_t* ids, void* errmsg);
 
 // Add chunk of data (same type as index quantization)
-void gpu_cagra_add_chunk(gpu_cagra_c index_c, const void* chunk_data, uint64_t chunk_count, const uint32_t* ids, void* errmsg);
+void gpu_cagra_add_chunk(gpu_cagra_c index_c, const void* chunk_data, uint64_t chunk_count, const int64_t* ids, void* errmsg);
 
 // Add chunk of data (from float, with on-the-fly quantization if needed)
-void gpu_cagra_add_chunk_float(gpu_cagra_c index_c, const float* chunk_data, uint64_t chunk_count, const uint32_t* ids, void* errmsg);
+void gpu_cagra_add_chunk_float(gpu_cagra_c index_c, const float* chunk_data, uint64_t chunk_count, const int64_t* ids, void* errmsg);
 
 // Trains the scalar quantizer (if T is 1-byte)
 void gpu_cagra_train_quantizer(gpu_cagra_c index_c, const float* train_data, uint64_t n_samples, void* errmsg);
@@ -84,7 +84,7 @@ void gpu_cagra_save(gpu_cagra_c index_c, const char* filename, void* errmsg);
 void gpu_cagra_save_dir(gpu_cagra_c index_c, const char* dir, void* errmsg);
 
 // Delete ID from index (soft delete via bitset)
-void gpu_cagra_delete_id(gpu_cagra_c index_c, uint32_t id, void* errmsg);
+void gpu_cagra_delete_id(gpu_cagra_c index_c, int64_t id, void* errmsg);
 
 // Load all components from a directory previously written by gpu_cagra_save_dir.
 // The index must have been created (e.g. via gpu_cagra_new_empty) and started before calling this.
@@ -116,7 +116,7 @@ gpu_cagra_search_res_t gpu_cagra_search_wait(gpu_cagra_c index_c, uint64_t job_i
 
 // Get results from result object
 
-void gpu_cagra_get_neighbors(gpu_cagra_result_c result_c, uint64_t total_elements, uint32_t* neighbors);
+void gpu_cagra_get_neighbors(gpu_cagra_result_c result_c, uint64_t total_elements, int64_t* neighbors);
 void gpu_cagra_get_distances(gpu_cagra_result_c result_c, uint64_t total_elements, float* distances);
 
 // Free result object
@@ -134,7 +134,7 @@ char* gpu_cagra_info(gpu_cagra_c index_c, void* errmsg);
 // Extend function
 // new_ids may be NULL to auto-assign sequential IDs starting from current index size
 void gpu_cagra_extend(gpu_cagra_c index_c, const void* additional_data, uint64_t num_vectors,
-                      const uint32_t* new_ids, void* errmsg);
+                      const int64_t* new_ids, void* errmsg);
 
 // Merge function
 gpu_cagra_c gpu_cagra_merge(gpu_cagra_c* indices_c, int num_indices, uint32_t nthread, const int* devices, int device_count, void* errmsg);
